Apply Scott's requested changes
