+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/directory/FDBTieredMergePolicy.java
@@ -1,0 +1,50 @@
+/*
+ * FDBTieredMergePolicy.java
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2015-2023 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.apple.foundationdb.record.lucene.directory;
+
+import org.apache.lucene.index.MergeTrigger;
+import org.apache.lucene.index.SegmentInfos;
+import org.apache.lucene.index.TieredMergePolicy;
+
+import java.io.IOException;
+
+class FDBTieredMergePolicy extends TieredMergePolicy {
+    final boolean shouldAutoMergeDuringCommit;
+
+    public FDBTieredMergePolicy(final boolean shouldAutoMergeDuringCommit) {
+        this.shouldAutoMergeDuringCommit = shouldAutoMergeDuringCommit;
+    }
+
+    boolean isAutoMergeDuringCommit(MergeTrigger mergeTrigger) {
+        return mergeTrigger == MergeTrigger.FULL_FLUSH ||
+               mergeTrigger == MergeTrigger.COMMIT;
+    }
+
+    @Override
+    public MergeSpecification findMerges(MergeTrigger mergeTrigger, SegmentInfos infos, MergeContext mergeContext) throws IOException {
+        if (!shouldAutoMergeDuringCommit && isAutoMergeDuringCommit(mergeTrigger)) {
+            // Here: skip it. The merge should be performed later by the user.
+            return null;
+        }
+        return super.findMerges(mergeTrigger, infos, mergeContext);
+    }
+}
+
